requirements.txt
Pillow==9.2.0
inference-cli>=0.9.9